Add third sensor and pump; change naming scheme
Adding additional sensors and pumps messed up the previous naming
convention, because it was pin-based and I had to change the position of
devices to fit them on the board.  Now sensors and pumps are named
according to their cable colours.  This results in combinations such as
"bro" (brown/red/orange) and "kkk" (black/black/black).